+++ docs/development/testing_guide.rst
@@ -115,7 +115,7 @@ short traceback reports, and run tests only if the test path contains
 
   python setup.py test -a "--maxfail=2 --tb=short -k 'plasma and not blob'"
 
-One may also simply run ``pytest`` as a shortcut from the command line.
+One may also run ``pytest`` from the command line.
 
 .. _testing-guidelines-running-tests-python:
 
@@ -338,16 +338,16 @@ other `~astropy.units.Quantity` objects to within a certain tolerance.
 Occasionally tests will be needed to make sure that a function will
 return the same value for different arguments (e.g., due to symmetry
 properties). PlasmaPy's `~plasmapy.utils` subpackage contains the
-`~plasmapy.utils.run_test` and
-`~plasmapy.utils.run_test_equivalent_calls` helper functions that can
+`~plasmapy.utils.pytest_helpers.run_test` and
+`~plasmapy.utils.pytest_helpers.run_test_equivalent_calls` helper functions that can
 generically perform many of these comparisons and checks.
 
-The `~plasmapy.utils.run_test` function can be used to check that a
-callable object returns the expected result, raises the expected
-exception, or issues the expected warning for different positional and
-keyword arguments.  This function is particularly useful when unit
-testing straightforward functions when you have a bunch of inputs and
-know the expected result.
+The `~plasmapy.utils.pytest_helpers.run_test` function can be used to
+check that a callable object returns the expected result, raises the
+expected exception, or issues the expected warning for different
+positional and keyword arguments. This function is particularly useful
+when unit testing straightforward functions when you have a bunch of
+inputs and know the expected result.
 
 Suppose that we want to test the trigonometric property that
 
@@ -355,24 +355,24 @@ Suppose that we want to test the trigonometric property that
 
   \sin(\theta) = \cos(\theta + \frac{\pi}{2}).
 
-We may use `~plasmapy.utils.run_test` as in the following example to
+We may use `~plasmapy.utils.pytest_helpers.run_test` as in the following example to
 check the case of :math:`\theta \equiv 0`.
 
 .. code-block:: python
 
   from numpy import sin, cos, pi
-  from plasmapy.utils import run_test
+  from plasmapy.utils.pytest_helpers import run_test
 
   def test_trigonometric_properties():
       run_test(func=sin, args=0, expected_outcome=cos(pi/2), atol=1e-16)
 
-We may use `pytest.mark.parametrize` with `~plasmapy.utils.run_test` to
-check multiple cases.  If `~plasmapy.utils.run_test` only receives one
-positional argument that is a `list` or `tuple`, then it will assume
-that `list` or `tuple` contains the `callable`, the positional
-arguments, the keyword arguments (which may be omitted), and the
-expected outcome (which may be the returned `object`, a warning, or an
-exception).
+We may use `pytest.mark.parametrize` with
+`~plasmapy.utils.pytest_helpers.run_test` to check multiple cases.  If
+`~plasmapy.utils.pytest_helpers.run_test` only receives one positional
+argument that is a `list` or `tuple`, then it will assume that `list`
+or `tuple` contains the `callable`, the positional arguments, the
+keyword arguments (which may be omitted), and the expected outcome
+(which may be the returned `object`, a warning, or an exception).
 
 .. code-block:: python
 
@@ -400,8 +400,8 @@ code.
       plasmapy.utils.run_test_equivalent_calls(cos, 1, -1)
 
 We may also use `pytest.mark.parametrize` with
-`~plasmapy.utils.run_test_equivalent_calls` to sequentially test
-multiple symmetry properties.
+`~plasmapy.utils.pytest_helpers.run_test_equivalent_calls` to
+sequentially test multiple symmetry properties.
 
 .. code-block:: python
 
@@ -413,10 +413,15 @@ This parametrized function will check that ``cos(1)`` is within
 ``1e-16`` of ``cos(-1)``, and that ``cos(pi/2)`` is within ``1e-16`` of
 ``sin(0)``.
 
-Please refer to the documentation for `~plasmapy.utils.run_test` and
-`~plasmapy.utils.run_test_equivalent_calls` to learn about the full
-capabilities of these pytest helper functions (including for testing
-functions that return `~astropy.units.Quantity` objects).
+Please refer to the documentation for
+`~plasmapy.utils.pytest_helpers.run_test` and
+`~plasmapy.utils.pytest_helpers.run_test_equivalent_calls` to learn
+about the full capabilities of these pytest helper functions (including
+for testing functions that return `~astropy.units.Quantity` objects).
+
+.. warning::
+    The API within `~plasmapy.utils.pytest_helpers` is not yet stable
+    and may change in the near future.
 
 .. _testing-guidelines-writing-tests-fixtures:
