+++ plasmapy/atomic/parsing.py
@@ -14,9 +14,8 @@ from plasmapy.atomic.elements import (
 from plasmapy.atomic.isotopes import _Isotopes
 from plasmapy.atomic.special_particles import _Particles, ParticleZoo
 
-from plasmapy.utils import (
-    roman
-)
+from plasmapy.utils import roman, pytest_helpers
+
 from plasmapy.atomic.exceptions import InvalidElementError, InvalidParticleError, AtomicWarning
 
 
@@ -433,13 +432,14 @@ def _parse_and_check_atomic_input(
     return nomenclature_dict
 
 
-def _call_string(arg: Union[str, Integral], kwargs: Dict = {}) -> str:
+def _particle_call_string(arg: Union[str, Integral], kwargs: Dict = {}) -> str:
     """
     Return a `str` that recreates the call to create a particular
-    particle from the input.
+    `~plasmapy.atomic.Particle` instance from the input.
     """
-    if kwargs != {}:
-        keyword_string = ", " + str(kwargs).strip(r"}{'").replace("'", "").replace(":", " =")
-    else:
-        keyword_string = ""
-    return f"Particle({repr(arg)}{keyword_string})"
+    return pytest_helpers.call_string(Particle, arg, kwargs)
+#    if kwargs != {}:
+#        keyword_string = ", " + str(kwargs).strip(r"}{'").replace("'", "").replace(":", " =")
+#    else:
+#        keyword_string = ""
+#    return f"Particle({repr(arg)}{keyword_string})"
