+++ plasmapy/atomic/tests/test_ionization_states.py
@@ -7,7 +7,7 @@ import itertools
 import numpy as np
 import astropy.units as u
 
-from plasmapy.utils import run_test
+from plasmapy.utils.pytest_helpers import run_test
 from plasmapy.atomic.exceptions import AtomicError, InvalidIsotopeError
 from plasmapy.atomic import (
     State,
