+++ plasmapy/utils/__init__.py
@@ -23,17 +23,3 @@ from .exceptions import (
     CouplingWarning,
     RelativityWarning,
 )
-
-from plasmapy.utils.pytest_helpers.pytest_helpers import (
-    run_test,
-    run_test_equivalent_calls,
-    call_string,
-    InconsistentTypeError,
-    UnexpectedResultError,
-    UnexpectedExceptionError,
-    RunTestError,
-    IncorrectResultError,
-    MissingExceptionError,
-    MissingWarningError,
-    assert_can_handle_nparray,
-)
